cleans up instrument.

+++ platforms/unity/Assets/Examples/Scripts/Bouncer.cs
@@ -6,11 +6,11 @@ using Barely;
 public class Bouncer : MonoBehaviour {
   public Instrument instrument;
 
-  public double rootPitch = 0.0f;
+  public double rootPitch = 0.0;
 
-  private readonly double[] MajorScale = new double[] { 0.0f, 2.0f, 4.0f, 5.0f, 7.0f, 9.0f, 11.0f };
+  private readonly double[] MajorScale = new double[] { 0.0, 2.0, 4.0, 5.0, 7.0, 9.0, 11.0 };
 
-  private readonly double NoteDuration = 0.05f;
+  private readonly double NoteDuration = 0.05;
 
   private int lastIndex = 0;
 
@@ -27,14 +27,14 @@ public class Bouncer : MonoBehaviour {
 
   private void OnCollisionEnter(Collision collision) {
     double octaveOffset = lastIndex / MajorScale.Length;
-    double pitch = rootPitch + octaveOffset + MajorScale[lastIndex++ % MajorScale.Length] / 12.0f;
-    double intensity = Mathf.Min(1.0f, 0.125f * collision.relativeVelocity.sqrMagnitude);
+    double pitch = rootPitch + octaveOffset + MajorScale[lastIndex++ % MajorScale.Length] / 12.0;
+    double intensity = (double)Mathf.Min(1.0f, 0.125f * collision.relativeVelocity.sqrMagnitude);
     StartCoroutine(PlayNote(pitch, intensity, NoteDuration));
   }
 
   IEnumerator PlayNote(double pitch, double intensity, double duration) {
     instrument.StartNote(pitch, intensity);
-    yield return new WaitForSeconds(duration);
+    yield return new WaitForSeconds((float)duration);
     instrument.StopNote(pitch);
   }
 }
